% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Add Custom Trait}
\alias{Add Custom Trait}
\alias{Add_Custom_Trait}
\title{Add Custom Trait}
\usage{
Add_Custom_Trait(
  RNAseq.data,
  Custom.traits,
  pairwise.distances,
  distance.metrics,
  bkgd.traits
)
}
\description{
Add a custom trait to an existing trait library
}
\author{
JJM van Steenbrugge
}
