% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Calc_TnA_redundancy}
\alias{Calc_TnA_redundancy}
\title{Calculates the number of overlapping traits and overlapping attributes for each pair of genomes}
\usage{
Calc_TnA_redundancy()
}
\description{
Calculates the number of overlapping traits and overlapping attributes for each pair of genomes
}
