% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Z_scores.R
\name{Calc_Z_scores}
\alias{Calc_Z_scores}
\title{Calculate Z scores of individual background distributions}
\usage{
Calc_Z_scores(bkgd.individual, metrics)
}
\arguments{
\item{bkgd.individual}{Named list containing (three) different background distributions.}

\item{metrics}{Named list containing possibly multiple functions for distance.}
}
\description{
Function to convert different random background distribution values of possible different distance metrics into Z_scores
}
