% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Create.Rank.Columns}
\alias{Create.Rank.Columns}
\title{Add a ranking collumn for each sample that holds the negative rank of transcripts
assigned to each gene, per bin. Normalized on the highest negative rank.}
\usage{
Create.Rank.Columns(RNAseq.table, RNAseq.features)
}
\description{
Add a ranking collumn for each sample that holds the negative rank of transcripts
assigned to each gene, per bin. Normalized on the highest negative rank.
}
\author{
JJM van Steenbrugge

BO Oyserman
}
