% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Create_Filtered_SBS_Matrix}
\alias{Create_Filtered_SBS_Matrix}
\title{Create_Filtered_SBS_Matrix}
\usage{
Create_Filtered_SBS_Matrix(
  trait.attributes.pruned,
  RNAseq.data,
  traits = c("M00171", "M00172")
)
}
\description{
Filters pruned trait attributes for attributes with a specific name. Then returns a sbs matrix of those
traits.
}
