% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Draw Expression}
\alias{Draw Expression}
\alias{Draw_Expression}
\title{Draw_Expression}
\usage{
Draw_Expression(trait, RNAseq.data, trait.attributes.pruned)
}
\arguments{
\item{trait}{character string with the name of a trait.}

\item{RNAseq.data}{Collection of multple components, include RNA seq data,
annotations, etc.}

\item{trait.attributes.pruned}{The full list of all pruned trait attributes.}
}
\description{
Creates a drawing of a trait (possibly with multiple different
trait-attributes) where all its composing genes are plotting individually
with their corresponding expression values.
}
\seealso{
\code{\link{Pre_process_input}} for the full list of parameters,
\url{https://github.com/Jorisvansteenbrugge/TcT/wiki/The-RNAseq.data-object}
for more information.
}
\author{
JJM van Steenbrugge
}
