% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Filter}
\alias{Filter}
\title{Remove data rows that have a stdev of zero.}
\usage{
Filter(RNAseq.table, sample.columns, filter.method)
}
\arguments{
\item{sample.columns}{a vector with the collumns that contain RNAseq data.}
}
\description{
Remove data rows that have a stdev of zero.
}
\author{
JJM van Steenbrugge
}
