% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Get_annotation_presence_absence}
\alias{Get_annotation_presence_absence}
\title{Calculate which KO is present for each Bin}
\usage{
Get_annotation_presence_absence(RNAseq.table, bins, annotation.db)
}
\arguments{
\item{bins}{a vector with all unique bins}
}
\description{
Calculate which KO is present for each Bin
}
\author{
JJM van Steenbrugge
}
