% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Get_KEGG_Sub_Modules}
\alias{Get_KEGG_Sub_Modules}
\title{Calculate the Disjunctive Normal Form (DNF) of Kegg modules. This only has to be re-run if the KEGG database is updated.
(Advanced function)}
\usage{
Get_KEGG_Sub_Modules()
}
\description{
Calculate the Disjunctive Normal Form (DNF) of Kegg modules. This only has to be re-run if the KEGG database is updated.
(Advanced function)
}
