% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Get_matrix_features}
\alias{Get_matrix_features}
\title{Automatically identify features based on column names}
\usage{
Get_matrix_features(RNAseq.table)
}
\value{
A list containing a vector of unique bins, a vector with the sample column
numbers, a vector containing the rank column numbers and a annotation_presence_absence matrix
}
\description{
Robust method to infer information from matrix column names
}
\author{
JJM van Steenbrugge
}
