% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Get_trait_presence_absence}
\alias{Get_trait_presence_absence}
\title{Determine for each genome whether a trait is present or absent}
\usage{
Get_trait_presence_absence(RNAseq.data)
}
\description{
Determine for each genome whether a trait is present or absent
}
\author{
JJM van Steenbrugge
}
