% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getMetricDist}
\alias{getMetricDist}
\title{Used by Plot_Pathway_genes internally}
\usage{
getMetricDist(metric, cat.genes, distance.metrics, RNAseq.data,
  bkgd.individual, bkgd.individual.Zscores)
}
\description{
Used by Plot_Pathway_genes internally
}
