% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wrapper.R
\name{GetTraitAttributes}
\alias{GetTraitAttributes}
\title{GetTraitAttributes}
\usage{
GetTraitAttributes(
  RNAseq.data = RNAseq.data,
  N = 1000,
  metrics = distance.metrics,
  p = 4,
  show_progres = TRUE
)
}
\arguments{
\item{N}{Number of iterations for the random background distributions}

\item{p}{number of CPU threads to use simultaneously}
}
\value{
trait.attributes.pruned
}
\description{
Wrapper of trait identification functions, including statistical validation
}
\examples{
GetTraitAttributes()
GetTraitAttributes(RNAseq.data, N = 1000, p = 4)
}
\author{
JJM van Steenbrugge
}
