% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Trait_Attributes.R
\name{Identify Trait Attributes}
\alias{Identify Trait Attributes}
\alias{Identify_Trait_Attributes}
\title{Identify Trait Attributes}
\usage{
Identify_Trait_Attributes(RNAseq_Annotated_Matrix, pairwise.distances,
  annotation.db, threads = 4)
}
\arguments{
\item{filepath}{RNAseq data file}

\item{distance.metrics}{Variable}
}
\description{
For each trait (or module) Attributes are calculated
}
\author{
JJM van Steenbrugge
}
