% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Traits.R
\name{Identify_Significance_Trait}
\alias{Identify_Significance_Trait}
\title{Identify_Significance_Trait}
\usage{
Identify_Significance_Trait(trait, RNAseq.data, distance.metrics, bkgd.traits)
}
\arguments{
\item{trait}{character string of the trait.}

\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{distance.metrics}{Named list containing distance functions as values.}

\item{bkgd.traits}{Random Background distributions of Traits.}
}
\description{
Gives a verdict wether a trait is significant as a whole.
}
\author{
JJM van Steenbrugge
}
