% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{Individual Random Background distributions}
\alias{Individual Random Background distributions}
\alias{Individual_Annotation_Background}
\title{Individual Random Background distributions}
\usage{
Individual_Annotation_Background(RNAseq.data, N = 1000, metrics = list(PC =
  PC, NRED = NRED))
}
\arguments{
\item{N}{Number of iterations}

\item{metrics}{Named list containing possibly multiple functions for distance
metrics}

\item{threads}{Number of cpu cores to be used}
}
\description{
A function that creates three different random background distributions
For each of them, two different genomes are sampled. Then for the first distribution,
from each of the two genomes, a random gene is selected. For the second distribution,
from each of the two genomes, a random gene that has an annotation is selected.
For the third distribution, an annotation is selected that is present in both genomes.
Then for each of the two genomes, a gene is selected with that annotation.
}
\author{
JJM van Steenbrugge
BO Oyserman
}
