% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{Individual_Annotation_Background}
\alias{Individual_Annotation_Background}
\title{Individual Random Background distributions}
\usage{
Individual_Annotation_Background(RNAseq.data, N = 1000, threads = 3,
  metrics = list(PC = PC, NRED = NRED))
}
\arguments{
\item{N}{Number of iterations}

\item{threads}{Number of cpu cores to be used}

\item{metrics}{Named list containing possibly multiple functions for distance
metrics}
}
\description{
Individual Random Background distributions
}
\author{
JJM van Steenbrugge
}
