% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Model_Module}
\alias{Model_Module}
\title{Model Module}
\usage{
Model_Module(RNAseq.data, trait.attributes, Model_Bin, Module_Names,
  Bin_Order, Yrange)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data,
annotations, etc.}

\item{Yrange}{It is optional to provide a Yrange}

\item{trait}{character string with the name of a trait.}
}
\description{
Creates a bar plot of pairwise comparisons between the traits of a model genome.
If a distance is less than the 99% quantile of the background distances for a module of similar size,
The bar is presented in dark grey.
}
\examples{
Model_Bin <- 39
\dontrun{
Bin_Order <- c(48,32,31,29,22,11,39,16,53,45,42,28,20,25,19,8,36,26,17)
Yrange <- c(-2.5,2.5)
Module_Names_Polymer_Metabolism <- c("M00001","M00307","HOX", "M00579","PHA","M00434","M00222", "PPK")
Module_Names_Purine_Metabolism <-c("M00048","M00049","M00050","M00546")
Module_Names_Pyrimidine_Metabolism <-c("M00051","M00052","M00053","M00046")
Module_Names_Nitrogen_Metabolism <- c("M00175","M00531","M00530","M00529","M00528","M00804")
Module_Names_CC_Metabolism <- c("M00001","M00002","M00003","M00307","M00009","M00010","M00011","M00004","M00006","M00007","M00580","M00005","M00008","M00308","M00633","M00309")
}
}
\seealso{
\code{\link{Pre_process_input}} for the full list of parameters,
\url{https://github.com/Jorisvansteenbrugge/TcT/wiki/The-RNAseq.data-object}
for more information.
}
\author{
BO Oyserman
}
