% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rcyto.R
\name{Network_Association_Rules}
\alias{Network_Association_Rules}
\title{Draw a network of association rules in Cytoscape}
\usage{
Network_Association_Rules(rules.dataframe, annotation.db, N = 200)
}
\arguments{
\item{rules.dataframe}{A dataframe containing association rules,
as calculated by \code{\link{Association_Rules}}.}

\item{annotation.db}{List containing a dictionary like structure with traits as
names and annotation as values.}

\item{N}{The number of association rules to visualise}
}
\description{
Association rules between trait-attributes can be visualized in
a directed graph network.
}
\author{
JJM van Steenbrugge
}
