% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Normalization}
\alias{Normalization}
\title{Normalize RNAseq data}
\usage{
Normalization(RNAseq.table, RNAseq.features, normalize.method)
}
\arguments{
\item{normalize.method}{variable that is either a string (default method) or a function to be applied
for normalization of RNAseq data.}
}
\description{
Normalize RNAseq data
}
\author{
JJM van Steenbrugge
}
