% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{NRED}
\alias{NRED}
\title{Calculates the Normalized Rank Euclidean Distance of two lines from
RNAseq.data$table}
\usage{
NRED(rowA, rowB, RNAseq.features)
}
\description{
Calculates the Normalized Rank Euclidean Distance of two lines from
RNAseq.data$table
}
