% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Pairwise_Annotation_Distance.R
\name{Pairwise comparison of Annotations}
\alias{Pairwise comparison of Annotations}
\alias{Calc_Pairwise_Annotation_Distance}
\title{Create an all versus all comparison for each annotation term}
\usage{
Calc_Pairwise_Annotation_Distance(RNAseq.data, annotation.db, distance.metrics,
  bkgd.individual.Zscores, show.progress = F)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{annotation.db}{List containing a dictionary like structure with trait as names
and annotation as values.}

\item{distance.metrics}{Named list containing distance functions as values}

\item{bkgd.individual.Zscores}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{show.progress}{Boolean indicating werther or not to show progress.}

\item{threads}{The number of cpu threads to run the function in parallel .}
}
\description{
For each annotation term, a comparison is done for each genome
versus all other genomes, 1 versus 1.
}
\author{
JJM van Steenbrugge
}
