% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{PC}
\alias{PC}
\title{Calculates the Pearson Correlation between two lines from RNAseq.data$table}
\usage{
PC(rowA, rowB, RNAseq.features)
}
\description{
Calculates the Pearson Correlation between two lines from RNAseq.data$table
}
