% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Background Individual Genes}
\alias{Plot Background Individual Genes}
\alias{Plot_Background_Individual_Genes}
\title{Plot_Background_Individual_Genes}
\usage{
Plot_Background_Individual_Genes(bkgd.individual.Zscores)
}
\description{
Plotting of each background distribution.
}
\author{
BO Oyserman
}
