% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R, R/Utility_2.R
\name{Plot Background Modules}
\alias{Plot Background Modules}
\alias{Plot_Background_Modules}
\title{Plot_Background_Modules}
\usage{
Plot_Background_Modules(bkgd.traits)

Plot_Background_Modules(bkgd.traits)
}
\arguments{
\item{bkgd.traits}{A collection of random background distributions of traits.
See \code{\link{Random_Trait_Background}}.}

\item{bkgd.traits}{A collection of random background distributions of traits.
See \code{\link{Random_Trait_Background}}.}
}
\description{
Plotting of all background distributions of different sizes
together. Each distribution is represented by a different colour. This plot
hints wether or not the background distribution is indeed random.

Plotting of all background distributions of different sizes
together. Each distribution is represented by a different colour. This plot
hints wether or not the background distribution is indeed random.
}
\author{
JJM van Steenbrugge

JJM van Steenbrugge
}
