% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Metric Comparison}
\alias{Plot Metric Comparison}
\alias{Plot_Metric_Comparison}
\title{Plot_Metric_Comparison}
\usage{
Plot_Metric_Comparison(bkgd.individual)
}
\description{
Plotting densities and significance of each distance metric
}
\author{
BO Oyserman
}
