% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Pathway genes}
\alias{Plot Pathway genes}
\alias{Plot_Pathway_genes}
\alias{Plot_Pathway_modules}
\title{Plot Pathway Genes}
\usage{
Plot_Pathway_genes(metric_name, distance.metrics, RNAseq.data,
  bkgd.individual, bkgd.individual.Zscores)

Plot_Pathway_modules()
}
\description{
(Figure 2b)

Currently no example, this is an expanded version of Figure 2b,
except at module level and not at pathway level.
}
