% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Trait Attribute Expression}
\alias{Plot Trait Attribute Expression}
\alias{Plot_Trait_Attribute_Expression}
\title{Plot Trait Attribute}
\usage{
Plot_Trait_Attribute_Expression(trait.attribute, trait.attributes.pruned,
  RNAseq.data)
}
\arguments{
\item{trait.attribute}{The name of the trait attribute (as character string)}

\item{RNAseq.data}{Collection of multple components, include RNA seq data,
annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{trait.attributes}{The full list of all trait attributes}
}
\description{
Plot the expression profile for all genomes expressing that
attribute. For each annotation a separate plot is created where the thick
black line represents the mean expression of all genomes with that annotation.
The dashed blue and red lines represent the mean + 0.5*sd and the mean -
0.5 * sd, respectively.
}
\author{
JJM van Steenbrugge
}
