% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Trait Attribute}
\alias{Plot Trait Attribute}
\alias{Plot_Trait_Attribute}
\title{Plot Trait Attribute}
\usage{
Plot_Trait_Attribute(trait.attribute, trait.attributes, RNAseq.data)
}
\arguments{
\item{trait.attribute}{The name of the trait attribute (as character string)}

\item{trait.attributes}{The full list of all trait attributes}

\item{RNAseq}{data}
}
\description{
Plot the expression profile for all genomes expressing that attribute.
For each annotation a separate plot is created where the thick black line represents
the mean expression of all genomes with that annotation. The dashed blue and red lines
represent the mean + 0.5*sd and the mean - 0.5 * sd, respectively
}
\author{
JJM van Steenbrugge
}
