% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot_Model_Module}
\alias{Plot_Model_Module}
\title{Plot_Model_Module}
\usage{
Plot_Model_Module(Model_Module_List, Model_Bin, Module_Names, margins)
}
\description{
Creates a bar plot of pairwise comparisons between the traits of a model genome.
If a distance is less than the 99% quantile of the background distances for a module of similar size,
The bar is presented in dark grey.
}
\seealso{
\code{\link{Pre_process_input}} for the full list of parameters,
\url{https://github.com/Jorisvansteenbrugge/TcT/wiki/The-RNAseq.data-object}
for more information.
}
\author{
BO Oyserman
}
