% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot_Redundancy_Traits}
\alias{Plot_Redundancy_Traits}
\title{Plot_Redundancy_Traits}
\usage{
Plot_Redundancy_Traits(RNAseq.data)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}
}
\description{
Plots a histogram of traits redundancy
}
\author{
JJM van Steenbrugge
}
