% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot_Shared_Attributes}
\alias{Plot_Shared_Attributes}
\title{Plot_Shared_Attributes}
\usage{
Plot_Shared_Attributes(trait.attributes.pruned, RNAseq.data)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data,}

\item{bins_of_interest}{A vectr of interesting bins}
}
\description{
plots the number of shared attributes between a set of bins and the rest of the genomes
}
