% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot_Trait_Expression}
\alias{Plot_Trait_Expression}
\title{Plot_Trait_Expression}
\usage{
Plot_Trait_Expression(trait, subset_genomes)
}
\description{
# Hacking categories because the order is known ----
  categories <- c(
    rep("P_Energy metabolism", 4),
    rep("P_Carbohydrate and lipid metabolism", 10),
    rep("P_Nucleotide and AA metabolism", 12),
    rep("P_Secondary metabolism", 2),
    rep("S_Energy metabolism", 2),
    rep("S_Genetic information processing", 10),
    rep("S_Environmental information processing", 9),
    rep("F_Metabolism", 2),
    rep("F_Environemntal information processing", 2),
    rep("F_Cellular processes", 1),
    rep("Si_Gene set", 4)
  )
  #
  cat.genes <- list()
}
\details{

}
