% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prune_Trait_Attributes.R
\name{Prune Trait Attributes}
\alias{Prune Trait Attributes}
\alias{Prune_Trait_Attributes}
\title{Prune Trait Attributes}
\usage{
Prune_Trait_Attributes(trait.attributes, bkgd.modules, annotation.db,
  p.threshold = 0.05, threads = 4)
}
\arguments{
\item{filepath}{RNAseq data file}

\item{distance.metrics}{Variable}
}
\description{
For each trait attribute, statistical significance is calculated
}
\author{
JJM van Steenbrugge
}
