% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Module_Background.R
\name{Random Background distributions of modules}
\alias{Random Background distributions of modules}
\alias{Module_Background}
\title{Generate a background distibution of composite Z scores}
\usage{
Module_Background(RNAseq.data, Z_scores, N, Z, metrics, threads = 2)
}
\arguments{
\item{N}{Number of genes to include in the random module}

\item{Z}{Number of iterations}

\item{metrics}{Named list containing possibly multiple functions for distance}

\item{threads}{Number of cpu cores to be used}

\item{range}{NOT SUPPORTED YET}
}
\description{
Generate a background distibution of composite Z scores
}
\author{
JJM van Steenbrugge
}
