% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Random_Trait_Background.R
\name{Random Background distributions of Traits}
\alias{Random Background distributions of Traits}
\alias{Random_Trait_Background}
\title{Generate a background distibution of composite Z scores}
\usage{
Random_Trait_Background(
  RNAseq.data,
  bkgd.individual.Zscores,
  N,
  metrics,
  Z,
  threads = 4
)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{N}{Number of genes to include in the random module}

\item{metrics}{Named list containing possibly multiple functions for distance}

\item{Z}{Number of iterations, if left empty the actual sizes of the modules in the database
will be used.}

\item{threads}{Number of cpu cores to be used}
}
\description{
Create a random background distribution of distances between genes in a random trait. When multiple distance metrics
are provided, a composite Z score is calculated. The background distribution is determined by first random sampling two genomes.
Between the two genomes the overlap in annotations is determined. Based on the overlap a trait
is artificially created containing N annotations. For each annotation in the trait, the distances are
calculated between Genome A and Genome B, as described in the previous section
}
\author{
JJM van Steenbrugge
BO Oyserman
}
