% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{Random.Annotated.Genes.bkgd}
\alias{Random.Annotated.Genes.bkgd}
\title{Background distribution of any two random genes with an annotation in two
random genomes}
\usage{
Random.Annotated.Genes.bkgd(RNAseq.data, metrics, N)
}
\description{
Background distribution of any two random genes with an annotation in two
random genomes
}
\author{
JJM van Steenbrugge
}
