% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Individual_Annotation_Background.R
\name{Random.Identical.Annotated.Genes.bkgd}
\alias{Random.Identical.Annotated.Genes.bkgd}
\title{Background distribution of two random random genes with the same
random annotation in two random genomes.}
\usage{
Random.Identical.Annotated.Genes.bkgd(RNAseq.data, metrics, N, random.genomes)
}
\arguments{
\item{random.genomes}{Allows reuse of this function in the module background function @see@seealso \code{\link{Random Background distributions of modules}}}
}
\description{
Background distribution of two random random genes with the same
random annotation in two random genomes.
}
\author{
JJM van Steenbrugge
}
