% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Shared traits}
\alias{Shared traits}
\alias{Shared_traits}
\alias{Plot_Shared_Attributes}
\title{Identify shared traits}
\usage{
Shared_traits(RNAseq.data, subset_genomes)

Plot_Shared_Attributes(trait.attributes.pruned, RNAseq.data,
  subset_genomes)
}
\description{
identify the shared traits between a subset of genomes

plot the shared traits between a subset of genomes
}
