% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Traitattributes_To_Sbsmatrix}
\alias{Traitattributes_To_Sbsmatrix}
\title{Traitattributes To Sbsmatrix}
\usage{
Traitattributes_To_Sbsmatrix(trait.attributes, genomes)
}
\arguments{
\item{trait.attributes}{The full list of all trait attributes}

\item{genomes}{A vector containing the names of all genomes present in the
sample.}
}
\description{
Convert a list with trait attributes to a matrix where each
collumn is a trait attribute and each row a genome. Presence or absence is
indicated with either a zero (0) or one (1).
}
\author{
JJM van Steenbrugge
}
