% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Z_scores.R
\name{Calc_Z_scores}
\alias{Calc_Z_scores}
\title{Calculate Z scores of individual background distributions}
\usage{
Calc_Z_scores(bkgd.individual, metrics)
}
\description{
Calculate Z scores of individual background distributions
}
