% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{getMetricDist}
\alias{getMetricDist}
\title{Get Metric Dist}
\usage{
getMetricDist(metric, cat.genes, distance.metrics, RNAseq.data,
  bkgd.individual, bkgd.individual.Zscores)
}
\arguments{
\item{metric}{either {'PC','NRED'}}

\item{cat.genes}{list("Carbon")}
}
\description{
Get Metric Dist
}
