% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Go Fish}
\alias{Go Fish}
\alias{Go_Fish}
\title{Identify whether an annotation term has a certain expression pattern in one
or more genomes}
\usage{
Go_Fish(RNAseq.data)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}
}
\description{
Identify whether a gene has a certain expression pattern in one or more genomes.
}
\author{
JJM van Steenbrugge
}
