% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GoFish.R
\name{Go Fish}
\alias{Go Fish}
\alias{Go_Fish}
\title{Identify whether an annotation term has a certain expression pattern in one
or more genomes}
\usage{
Go_Fish(annotation, RNAseq.data, pattern)
}
\arguments{
\item{annotation}{character string of the Annotation term}

\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc. See \code{\link{Pre_process_input}} for the full list.}

\item{pattern}{character string of the kind of pattern to look for. The string
should be one of the following: {'increase', 'decrease', 'peak', 'vale}}
}
\description{
Identify whether a gene has a certain expression pattern in one or more genomes.
}
\author{
JJM van Steenbrugge
}
