% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Identify_Trait_Attributes.R
\name{Identify Trait Attributes}
\alias{Identify Trait Attributes}
\alias{Identify_Trait_Attributes}
\title{Identify Trait Attributes}
\usage{
Identify_Trait_Attributes(
  RNAseq.data,
  pairwise.distances,
  annotation.db,
  threads = 4
)
}
\arguments{
\item{RNAseq.data}{Collection of multple components, include RNA seq data, annotations, etc.
See \code{\link{Pre_process_input}} for the full list.}

\item{pairwise.distances}{Named list containing possibly multiple functions for distance}

\item{annotation.db}{List containing a dictionary like structure with trait as names
and annotation as values.}
}
\description{
For each trait (or module) Attributes are calculated
}
\author{
JJM van Steenbrugge
BO Oyserman
}
