% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Normalization}
\alias{Normalization}
\title{Normalize RNAseq data}
\usage{
Normalization(
  RNAseq.table,
  RNAseq.features,
  normalize.method,
  normalization.features
)
}
\arguments{
\item{normalize.method}{variable that is either a function to be applied or
the default method for normalization of RNAseq data.}
}
\description{
Normalize RNAseq data
}
\author{
JJM van Steenbrugge
}
