% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Calc_Pairwise_Annotation_Distance.R
\name{Pairwise comparison of Annotations}
\alias{Pairwise comparison of Annotations}
\alias{Calc_Pairwise_Annotation_Distance}
\title{Create an all versus all comparison for each annotation term}
\usage{
Calc_Pairwise_Annotation_Distance(RNAseq.data, annotation.db, distance.metrics,
  bkgd.individual.Zscores, show.progress = F)
}
\arguments{
\item{distance.metrics}{Variable}

\item{filepath}{RNAseq data file}
}
\description{
For each annotation term, a comparison is done for each genome
versus all other genomes, 1 versus 1.
}
\author{
JJM van Steenbrugge
}
