% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Background Individual Genes}
\alias{Plot Background Individual Genes}
\alias{Plot_Background_Individual_Genes}
\title{Plot_Background_Individual_Genes}
\usage{
Plot_Background_Individual_Genes(bkgd.individual.Zscores)
}
\arguments{
\item{bkgd.individual.Zscores}{A list containing (composite) Zscores based on
a random background distribution.
See \code{\link{Calc_Z_scores}} for more information.}
}
\description{
Plotting of each background distribution (Figure 2A.
}
\author{
BO Oyserman
}
