% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Background Modules}
\alias{Plot Background Modules}
\alias{Plot_Background_Modules}
\title{Plot_Background_Modules}
\usage{
Plot_Background_Modules(bkgd.modules)
}
\description{
Plotting of each background distribution.
}
\author{
JJM van Steenbrugge
}
