% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utility.R
\name{Plot Metric comparison}
\alias{Plot Metric comparison}
\alias{Plot_Metric_Comparison}
\title{Plot_Metric_Comparison}
\usage{
Plot_Metric_Comparison(bkgd.individual)
}
\arguments{
\item{bkgd.individual.Zscores}{A list containing (composite) Zscores based on
a random background distribution.
See \code{\link{Calc_Z_scores}} for more information.}
}
\description{
A comparison of the two distance metrics used, with p-values plotted. Two comparisons
are made for each distance metric: 1) The distances between random genes and random annotated genes 2)
The distances between random annotated genes and random genes with the same annotation.
}
\author{
BO Oyserman
}
