% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{Pre_process_input}
\alias{Pre_process_input}
\title{Pre-processing of RNAseq data}
\usage{
Pre_process_input(filepath, normalize.method = FALSE,
  filter.method = "stdev")
}
\arguments{
\item{filepath}{RNAseq data file}

\item{normalize.method}{Variable}
}
\description{
Wrapper of preprocess functions, including: Identifying Matrix features
, Normalization, Filtering for non informative lines
}
\author{
JJM van Steenbrugge
}
